Modularized views.py and added tests
