+++ forum/views/__init__.py
@@ -1,0 +1,11 @@
+from .auth_views import *
+from .comments_views import *
+from .course_views import *
+from .notification_views import * 
+from .post_views import *
+from .profile_views import *
+from .save_posts_views import *
+from .search_views import *
+from .solution_views import *
+from .updates_views import *
+from .utils import *
